requirements.txt
astroid==2.2.4
autopep8==1.4.3
boto3==1.9.94
botocore==1.12.94
certifi==2018.11.29
chardet==3.0.4
docutils==0.14
idna==2.8
imgurpython==1.1.7
isort==4.3.11
jmespath==0.9.3
lazy-object-proxy==1.3.1
mccabe==0.6.1
numpy==1.16.1
Pillow==5.4.1
pycodestyle==2.5.0
pylint==2.3.1
python-dateutil==2.8.0
requests==2.21.0
s3transfer==0.2.0
six==1.12.0
typed-ast==1.3.1
urllib3==1.24.1
wrapt==1.11.1
